Updated Contributions.md
Updated the file such that the file paths to the screenshots are directed at the main branch. When the pull request is made, the screenshots won't appear in the Contributions.md on GitHub, but they will once it is merged to main (because the screenshots do not yet exist on the main branch).