Added Upstream Power Splits
This is an initial/interim fix to Issue #22. Instead of multiplying by the architecture matrices while computing the power available (which only works for architectures without any series connections), the user must provide 'Upstream' Operational matrices. These matrices are the same size as the other operational matrices, except the values must be filled in with the Upstream Power Splits. The upstream power split, P_{ij}, is defined as the power supplied to source 'i' from source 'j' divided by the total output power of source 'j'. See CreatePropArch for examples of these matrices.