Created UpstreamSplit
Created a function to compute the upstream power splits based on the power required at the upstream/downstream sources and the interdependency matrices. The upstream power splits, P_ij, are defined as the fraction of 'the power supplied by source j to source i' to 'the power output by source j'. This function will be implemented to help compute the power available for propulsion architectures containing power sources in series.