Updated PropulsionSizing
This is a temporary fix to account for the fact that the engine must produce the thrust given power from both itself and an electric motor. The thrust to be produced by the engine and electric motor is added together and the power provided by the electric motor is accounted for in the call to TurbofanNonlinearsizing.