The new BADA equation with calibration factor
The new BADA equation with calibration factor c for the SLS thrust of hybrid and conventional engine. The SLS thrust of the conventional engine is calculated by non-electrification case. And the SLS thrust of the hybrid engine is calculated by the SLS thrust of the conventional engine minus the electricload in takeoff.