Created PowerSupplementCheck
This code is a follow-up to code created previously that checks for parallel connections in the propulsion architecture and indicates how much power an electric motor is supplying to a gas-turbine engine. Rather than only checking for the previously described connection, this new code checks for both series connections (which extract power from a gas-turbine engine and supply it to an electric generator) and parallel connections (in which electric motors supplement power to a gas-turbine engine that is turning a propeller), and accordingly adjusts how much power is being supplied/siphoned from the gas-turbine engine.