Updated ResizeBattery
Fixed Bug: if the aircraft was not carrying a battery, then the battery resize function would return an empty vector, which would make MTOW empty in off-design analyses. Now, if there is no battery in the aircraft, the battery resize function returns a zero to show that there is no battery onboard.