Updated PropArchConnections
Added a logical to check that the driving and helping components actually exist. There is a chance that the FOR loop is initiated even though components are not in parallel (for example, if a redundant system is included). If the driving and helping components don't exist, nothing happens. If they do exist, it is noted in the structure that remembers the parallel connections.